+++ lib/flags.ts
@@ -1,3 +1,7 @@
+// `dank serve` will print http access logs to console
+export const isLogHttp = () =>
+    process.env.LOG_HTTP === 'true' || process.argv.includes('--log-http')
+
 // `dank serve` will pre-bundle and use service worker
 export const isPreviewBuild = () =>
     process.env.PREVIEW === 'true' || process.argv.includes('--preview')